Clean up the code/comments to be more idiomatic.
Now that I think about it more, I'm really unsure if
_glfwWin.oldMouseLockValid should be reset to false in
glfwDisable(GLFW_MOUSE_CURSOR). I'm just not sure of the
effect/usefulness of it, and it's not nice to have code that you don't
really understand or if it has no effect.

I initially added the code to reset oldMouseLockValid to false for both
glfwEnable and glfwDisable(GLFW_MOUSE_CURSOR) on the rationale that if
the user explicitly calls those functions, it doesn't make sense for
GLFW to automatically change the mouse lock to an older value.

But is there ever a case where the cursor is visible, then the window
loses focus/iconifies and the cursor is automatically made _invisible_
by GLFW. Then user calls glfwDisable(GLFW_MOUSE_CURSOR) and this should
make sure the cursor remains invisible after it regains focus (rather
than automatically being reset to visible as it was originally). I
don't think this happens because GLFW never does visible->invisible
cursor change, only the other way around.