Added an #ifdef block around Win32-specific code in enable.c.
This is a quick baseline solution that should make it compile. I'm open
for ideas for a better solution that doesn't require an #ifdef.

But the problem is, as the internal API is structured now, no
platform-specific code is ever called when user calls
glfwEnable(GLFW_MOUSE_CURSOR) while the cursor is made temporarily
visible by GLFW (in `win32/win32_window.c`). So as far as I can tell,
the only way to avoid this #ifdef is to add some sort of call that
happens in enableMouseCursor() when _glfwWin.mouseLock is == false. But
that seems to be way more verbosity for such a simple thing, I'm not
sure if it's justified.